deleted history.push and refactored the code so when I add a enw user it re-renders the app and shows it
