+++ friends/src/components/NewFriend.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import axios from 'axios';
 import { Link } from 'react-router-dom';
 
 class NewFriend extends Component {
@@ -9,7 +8,6 @@ class NewFriend extends Component {
       this.handleNameChange = this.handleNameChange.bind(this);
       this.handleAgeChange = this.handleAgeChange.bind(this);
       this.handleEmailChange = this.handleEmailChange.bind(this);
-      this.handleSubmit = this.handleSubmit.bind(this);
 
       this.state = {
           name: '',
@@ -30,39 +28,11 @@ class NewFriend extends Component {
         this.setState({email: e.target.value});
     }
 
-    handleSubmit(e) {
-        const self = this;
-        const random = Math.floor((Math.random() * 1000000000000000) + 1);
-
-        e.preventDefault();
-        axios.post('http://localhost:5000/friends', {
-            id: random,
-            name: self.state.name,
-            age: self.state.age,
-            email: self.state.email
-        })
-        .then(function (response) {
-            console.log(response);
-
-            self.setState({
-                name: '',
-                age: 0,
-                email: ''
-            });
-
-            
-            self.props.history.push('/')
-        })
-        .catch(function (error) {
-            console.log(error);
-        });
-    }
-
     render() {
         return (
             <div>
                 <Link to='/'>Home</Link>
-                <form onSubmit={this.handleSubmit}>
+                <form onSubmit={(e) => this.props.handleSubmit(e, this.state.name, this.state.age, this.state.email)}>
                     <label>
                         Name:
                         <input type="text" value={this.state.name} onChange={this.handleNameChange} />
